+++ start/src/App.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
   import { FIconCross } from '@fighting-design/fighting-icon'
+
+  const goBack = () => {
+    console.log('go back')
+  }
 </script>
 
 <template>
-  <f-avatar src="https://tianyuhao.cn/images/auto/my.jpg" background="red">
-    <template #icon>
-      <f-icon-cross />
-    </template>
-  </f-avatar>
+  <f-page-header title="标题" @back="goBack" />
 </template>